Minor nodeletctl code reorg
